refactor(cli): classify files
