+++ packages/cli/core/plugins/template/attrs/index.js
@@ -1,0 +1,7 @@
+exports = module.exports = function parseAttrs () {
+  [
+    './bindClass',
+    './bindStyle',
+    './src'
+  ].map(v => require(v).call(this));
+};
