+++ packages/cli/core/plugins/template/directives/index.js
@@ -1,0 +1,9 @@
+exports = module.exports = function parseDirectives () {
+  [
+    './additions',
+    './bind',
+    './for',
+    './model',
+    './on'
+  ].map(v => require(v).call(this));
+};
