+++ packages/cli/core/plugins/template/directives/model.js
@@ -88,4 +88,3 @@ exports = module.exports = function () {
     }
   });
 };
-
