+++ packages/cli/core/plugins/template/parse.js
@@ -1,13 +1,5 @@
 const htmlparser = require('htmlparser2');
-const paramsDetect = require('./../../ast/paramsDetect');
-const vueWithTransform = require('vue-template-es2015-compiler');
 const tools = require('../../util/tools');
-
-
-const forAliasRE = /([^]*?)\s+(?:in|of)\s+([^]*)/;
-const forIteratorRE = /,([^,\}\]]*)(?:,([^,\}\]]*))?$/;
-const stripParensRE = /^\(|\)$/g;
-const variableRE = /^\s*[a-zA-Z\$_][a-zA-Z\d_]*\s*$/;
 const onRE = /^@|^v-on:/;
 const bindRE = /^:|^v-bind:/;
 const modifierRE = /\.[^.]+/g;
@@ -41,173 +33,7 @@ const parseModifiers = (name = '') => {
   return ret;
 };
 
-
-/**
- * parsse handler AST
- * @param {String} expr   function expression, e.g. doSomething(a,b,c); item++;
- * @return {Object}       AST result
- */
-const parseHandlerProxy = (expr, scope) => {
-
-  let injectParams = [];
-  let handlerExpr = expr;
-  let functionName = 'proxyHandler';
-
-  if (/^\w+$/.test(expr)) {  //   @tap="doSomething"
-    injectParams.push('$event');
-    handlerExpr += '($event)';
-    functionName = 'proxyHandlerWithEvent';
-  } else {
-    let detected;
-    try {
-      detected = paramsDetect(handlerExpr);
-    } catch (e) {
-      throw new Error(`Can not parse "${handlerExpr}"`);
-    }
-
-    Object.keys(detected).forEach(d => {
-      if (scope && !detected[d].callable && scope.declared.indexOf(d) !== -1) {
-        injectParams.push(d);
-      }
-    });
-
-    if (detected.$event) {
-      injectParams.push('$event');
-      functionName = 'proxyHandlerWithEvent';
-    }
-  }
-
-
-  let proxy = `function ${functionName} (${injectParams.join(', ')}) {
-    with (this) {
-      return (function () {
-        ${handlerExpr}
-      })();
-    }
-  }`;
-
-  proxy = vueWithTransform(proxy);  // transform this
-  proxy = proxy.replace('var _h=_vm.$createElement;var _c=_vm._self._c||_h;', ''); // removed unused vue code;
-  return {
-    proxy: proxy,
-    params: injectParams
-  };
-};
-/**
- * parse event handler
- * @param  {String} key   event key, e.g. tap
- * @param  {String} value event value, e.g. doSomething(item)
- * @return {Object}       parse result, e.g. {type: "bind:tap", name: "doSomething", params: ["item"]}
- */
-const parseHandler = (key = '', value = '', modifiers = {}, scope) => {
-  let handler = '';
-  let type = '';
-  let info;
-  info = parseHandlerProxy(value.trim(), scope);
-
-  if (key === 'click')
-    key = 'tap';
-  if (modifiers.capture) {
-    type = 'capture-';
-  }
-  type = type + (modifiers.stop ? 'catch' : 'bind') + key;
-  return {
-    event: key,
-    type: type,
-    params: info.params,
-    proxy: info.proxy
-  };
-};
-
 exports = module.exports = function () {
-
-
-  this.register('template-parse-ast-attr-v-bind', function parseAstBind (item, name, value, modifiers, scope) {
-    return {
-      name: name,
-      prop: name.replace(bindRE, ''),
-      value: value,
-      expr: `{{ ${value} }}`
-    };
-  });
-
-  this.register('template-parse-ast-attr-v-on', function parseAstOn (item, evt, handler, modifiers, scope) {
-    evt = evt.replace(onRE, '');
-    let info = parseHandler(evt, handler, modifiers, scope);
-    let parsed = {};
-
-    info.params.forEach((p, i) => {
-      let paramAttr = 'data-wpy' + info.event.toLowerCase() + '-' + String.fromCharCode(97 + i);
-      if (paramAttr.length > 31) {
-        this.logger.warn(`Function name is too long, it may cause an Error. "${info.handler}"`);
-      }
-      parsed[paramAttr] = `{{ ${p} }}`;
-    });
-    parsed[info.type] = '_proxy';
-
-    info.parsed = parsed;
-    return info;
-  });
-
-  const ATTR_HANDLERS = {
-    'v-for': ({item, name, expr, parentScope}) => {
-      let res = {};
-      let scope = {};
-      let inMatch = expr.match(forAliasRE);
-      let variableMatch = expr.match(variableRE);
-      if (variableMatch) {
-        // e.g: v-for="items"
-        res.alias = 'item';
-        res.for = variableMatch[0].trim();
-        scope.for = res.for;
-        scope.declared = [];
-      }
-
-      if (inMatch) {
-        scope.declared = scope.declared || [];
-        res.for = inMatch[2].trim();
-        let alias = inMatch[1].trim().replace(stripParensRE, '');
-        let iteratorMatch = alias.match(forIteratorRE);
-        if (iteratorMatch) {
-          res.alias = alias.replace(forIteratorRE, '').trim();
-          scope.declared.push(res.alias);
-          scope.alias = res.alias;
-          res.iterator1 = iteratorMatch[1].trim();
-          scope.iterator1 = res.iterator1;
-          scope.declared.push(res.iterator1);
-          if (iteratorMatch[2]) {
-            res.iterator2 = iteratorMatch[2].trim();
-            scope.iterator2 = res.iterator2;
-            scope.declared.push(res.iterator2);
-          }
-        } else {
-          res.alias = alias;
-          scope.alias = alias;
-          scope.declared.push(alias);
-        }
-      }
-      if (parentScope)
-        scope.parent = parentScope;
-      return {
-        scope: scope,
-        attrs: {
-          'wx:for': `{{ ${res.for} }}`,
-          'wx:for-index': `${res.iterator1 || 'index'}`,
-          'wx:for-item': `${res.alias || 'item'}`,
-          'wx:key': `${res.iterator2 || res.iterator1 || 'index'}`
-        }
-      };
-    },
-    'v-show': ({item, name, expr}) => ({attrs: { hidden: `{{ !(${expr}) }}` }}),
-    'v-if': ({item, name, expr}) => ({attrs: { 'wx:if': `{{ ${expr} }}` }}),
-    'v-else-if': ({item, name, expr}) => ({attrs: { 'wx:elif': `{{ ${expr} }}` }}),
-    'v-else': ({item, name, expr}) => ({attrs: { 'wx:else': true }})
-  };
-
-  for (let name in ATTR_HANDLERS) {
-    this.register('template-parse-ast-attr-' + name, ATTR_HANDLERS[name]);
-  }
-
 
   this.register('template-parse-ast-attr', function parseAstAttr (item, scope, rel, ctx) {
     let attrs = item.attribs;
